dataProvider function part moved from the LoginTest.class to TestBase.class and verified working
