+++ src/main/java/base/TestBase.java
@@ -9,6 +9,7 @@ import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.DataProvider;
 
 import io.github.bonigarcia.wdm.WebDriverManager;
 
@@ -60,5 +61,17 @@ public class TestBase {
 			driver.quit();
 			driver=null;  
 			System.out.println("Tear down successful");		
+		}
+		
+		
+		
+		//data driven testing using the testNG annotation
+		@DataProvider(name = "userCredentialsFromInternalArray")
+		public Object[][] createData1() {
+			return new Object[][] {
+				{ "admin", "useradmin1"},
+				{ "admin1", "useradmin"},
+				{ "admin", "useradmin"}  
+			};
 		}
 }
