+++ test-output/Default suite/Default test.html
@@ -57,9 +57,9 @@ function toggleAllBoxes() {
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>3/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Thu May 08 19:56:43 IST 2025</td>
+<td>Started on:</td><td>Thu May 08 21:27:26 IST 2025</td>
 </tr>
-<tr><td>Total time:</td><td>51 seconds (51516 ms)</td>
+<tr><td>Total time:</td><td>31 seconds (31311 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -76,11 +76,6 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='tests.LoginTest.LoginFunc()'><b>LoginFunc</b><br>Test class: tests.LoginTest<br>Parameters: admin1, useradmin</td>
-<td></td>
-<td>1</td>
-<td>tests.LoginTest@1e800aaa</td></tr>
-<tr>
 <td title='tests.LoginTest.LoginFunc()'><b>LoginFunc</b><br>Test class: tests.LoginTest<br>Parameters: admin, useradmin1</td>
 <td></td>
 <td>2</td>
@@ -88,7 +83,12 @@ function toggleAllBoxes() {
 <tr>
 <td title='tests.LoginTest.LoginFunc()'><b>LoginFunc</b><br>Test class: tests.LoginTest<br>Parameters: admin, useradmin</td>
 <td></td>
-<td>4</td>
+<td>6</td>
+<td>tests.LoginTest@1e800aaa</td></tr>
+<tr>
+<td title='tests.LoginTest.LoginFunc()'><b>LoginFunc</b><br>Test class: tests.LoginTest<br>Parameters: admin1, useradmin</td>
+<td></td>
+<td>1</td>
 <td>tests.LoginTest@1e800aaa</td></tr>
 </table><p>
 </body>
