+++ test-output/Default suite/Default test.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-DCTEAPV" failures="0" tests="3" name="Default test" time="51.516" errors="0" timestamp="2025-05-08T19:57:35 IST">
-  <testcase classname="tests.LoginTest" name="LoginFunc" time="2.749"/>
-  <testcase classname="tests.LoginTest" name="LoginFunc" time="1.42"/>
-  <testcase classname="tests.LoginTest" name="LoginFunc" time="4.535"/>
+<testsuite ignored="0" hostname="DESKTOP-DCTEAPV" failures="0" tests="3" name="Default test" time="31.311" errors="0" timestamp="2025-05-08T21:27:58 IST">
+  <testcase classname="tests.LoginTest" name="LoginFunc" time="2.46"/>
+  <testcase classname="tests.LoginTest" name="LoginFunc" time="1.895"/>
+  <testcase classname="tests.LoginTest" name="LoginFunc" time="6.315"/>
 </testsuite> <!-- Default test -->
