+++ test-output/index.html
@@ -50,7 +50,7 @@
               <ul>
                 <li>
                   <a href="#" panel-name="test-xml-Default_suite" class="navigator-link ">
-                    <span>C:\Users\LENOVO\AppData\Local\Temp\testng-eclipse--1401403361\testng-customsuite.xml</span>
+                    <span>C:\Users\LENOVO\AppData\Local\Temp\testng-eclipse-1863370536\testng-customsuite.xml</span>
                   </a> <!-- navigator-link  -->
                 </li>
                 <li>
@@ -103,17 +103,17 @@
                     <div class="method-list-content passed suite-Default_suite">
                       <span>
                         <img src="passed.png" width="3%"/>
-                        <a href="#" panel-name="suite-Default_suite" title="tests.LoginTest" class="method navigator-link" hash-for-method="LoginFunc(admin1, useradmin)">LoginFunc(admin1, useradmin)</a> <!-- method navigator-link -->
-                      </span>
-                      <br/>
-                      <span>
-                        <img src="passed.png" width="3%"/>
                         <a href="#" panel-name="suite-Default_suite" title="tests.LoginTest" class="method navigator-link" hash-for-method="LoginFunc(admin, useradmin1)">LoginFunc(admin, useradmin1)</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                       <span>
                         <img src="passed.png" width="3%"/>
                         <a href="#" panel-name="suite-Default_suite" title="tests.LoginTest" class="method navigator-link" hash-for-method="LoginFunc(admin, useradmin)">LoginFunc(admin, useradmin)</a> <!-- method navigator-link -->
+                      </span>
+                      <br/>
+                      <span>
+                        <img src="passed.png" width="3%"/>
+                        <a href="#" panel-name="suite-Default_suite" title="tests.LoginTest" class="method navigator-link" hash-for-method="LoginFunc(admin1, useradmin)">LoginFunc(admin1, useradmin)</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                     </div> <!-- method-list-content passed suite-Default_suite -->
@@ -136,14 +136,6 @@
             <div class="main-panel-content rounded-window-bottom">
               <div class="method">
                 <div class="method-content">
-                  <a name="LoginFunc(admin1, useradmin)">
-                  </a> <!-- LoginFunc(admin1, useradmin) -->
-                  <span class="method-name">LoginFunc</span>
-                  <span class="parameters">(admin1,useradmin)</span>
-                </div> <!-- method-content -->
-              </div> <!-- method -->
-              <div class="method">
-                <div class="method-content">
                   <a name="LoginFunc(admin, useradmin1)">
                   </a> <!-- LoginFunc(admin, useradmin1) -->
                   <span class="method-name">LoginFunc</span>
@@ -158,12 +150,20 @@
                   <span class="parameters">(admin,useradmin)</span>
                 </div> <!-- method-content -->
               </div> <!-- method -->
+              <div class="method">
+                <div class="method-content">
+                  <a name="LoginFunc(admin1, useradmin)">
+                  </a> <!-- LoginFunc(admin1, useradmin) -->
+                  <span class="method-name">LoginFunc</span>
+                  <span class="parameters">(admin1,useradmin)</span>
+                </div> <!-- method-content -->
+              </div> <!-- method -->
             </div> <!-- main-panel-content rounded-window-bottom -->
           </div> <!-- suite-Default_suite-class-passed -->
         </div> <!-- panel Default_suite -->
         <div panel-name="test-xml-Default_suite" class="panel">
           <div class="main-panel-header rounded-window-top">
-            <span class="header-content">C:\Users\LENOVO\AppData\Local\Temp\testng-eclipse--1401403361\testng-customsuite.xml</span>
+            <span class="header-content">C:\Users\LENOVO\AppData\Local\Temp\testng-eclipse-1863370536\testng-customsuite.xml</span>
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <pre>
@@ -216,20 +216,20 @@ data.addRows(3);
 data.setCell(0, 0, 0)
 data.setCell(0, 1, 'LoginFunc')
 data.setCell(0, 2, 'tests.LoginTest')
-data.setCell(0, 3, 4535);
+data.setCell(0, 3, 6315);
 data.setCell(1, 0, 1)
 data.setCell(1, 1, 'LoginFunc')
 data.setCell(1, 2, 'tests.LoginTest')
-data.setCell(1, 3, 2749);
+data.setCell(1, 3, 2460);
 data.setCell(2, 0, 2)
 data.setCell(2, 1, 'LoginFunc')
 data.setCell(2, 2, 'tests.LoginTest')
-data.setCell(2, 3, 1420);
+data.setCell(2, 3, 1895);
 window.suiteTableData['Default_suite']= { tableData: data, tableDiv: 'times-div-Default_suite'}
 return data;
 }
               </script>
-              <span class="suite-total-time">Total running time: 8 seconds</span>
+              <span class="suite-total-time">Total running time: 10 seconds</span>
               <div id="times-div-Default_suite">
               </div> <!-- times-div-Default_suite -->
             </div> <!-- times-div -->
@@ -262,35 +262,35 @@ return data;
               </div> <!-- configuration-method before -->
               <div class="test-method">
                 <span class="method-name">LoginFunc(admin, useradmin1)</span>
-                <span class="method-start">33511 ms</span>
+                <span class="method-start">13608 ms</span>
               </div> <!-- test-method -->
               <div class="configuration-method after">
                 <span class="method-name">tearDown</span>
-                <span class="method-start">36270 ms</span>
+                <span class="method-start">16078 ms</span>
               </div> <!-- configuration-method after -->
               <div class="configuration-method before">
                 <span class="method-name">setUp</span>
-                <span class="method-start">36917 ms</span>
+                <span class="method-start">17133 ms</span>
               </div> <!-- configuration-method before -->
               <div class="test-method">
                 <span class="method-name">LoginFunc(admin1, useradmin)</span>
-                <span class="method-start">41351 ms</span>
+                <span class="method-start">20369 ms</span>
               </div> <!-- test-method -->
               <div class="configuration-method after">
                 <span class="method-name">tearDown</span>
-                <span class="method-start">42794 ms</span>
+                <span class="method-start">22268 ms</span>
               </div> <!-- configuration-method after -->
               <div class="configuration-method before">
                 <span class="method-name">setUp</span>
-                <span class="method-start">43362 ms</span>
+                <span class="method-start">22625 ms</span>
               </div> <!-- configuration-method before -->
               <div class="test-method">
                 <span class="method-name">LoginFunc(admin, useradmin)</span>
-                <span class="method-start">45593 ms</span>
+                <span class="method-start">24224 ms</span>
               </div> <!-- test-method -->
               <div class="configuration-method after">
                 <span class="method-name">tearDown</span>
-                <span class="method-start">50137 ms</span>
+                <span class="method-start">30542 ms</span>
               </div> <!-- configuration-method after -->
           </div> <!-- main-panel-content rounded-window-bottom -->
         </div> <!-- panel -->
