+++ test-output/junitreports/TEST-tests.LoginTest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-DCTEAPV" failures="0" tests="3" name="tests.LoginTest" time="50.587" errors="0" timestamp="2025-05-08T19:57:35 IST" skipped="0">
-  <testcase classname="tests.LoginTest" name="LoginFunc" time="35.492"/>
-  <testcase classname="tests.LoginTest" name="LoginFunc" time="5.481"/>
-  <testcase classname="tests.LoginTest" name="LoginFunc" time="9.614"/>
+<testsuite hostname="DESKTOP-DCTEAPV" failures="0" tests="3" name="tests.LoginTest" time="31.087" errors="0" timestamp="2025-05-08T21:27:59 IST" skipped="0">
+  <testcase classname="tests.LoginTest" name="LoginFunc" time="17.114"/>
+  <testcase classname="tests.LoginTest" name="LoginFunc" time="9.907"/>
+  <testcase classname="tests.LoginTest" name="LoginFunc" time="4.066"/>
 </testsuite> <!-- tests.LoginTest -->
