+++ test-output/testng-results.xml
@@ -2,16 +2,16 @@
 <testng-results ignored="0" total="3" passed="3" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2025-05-08T19:56:43 IST" name="Default suite" finished-at="2025-05-08T19:57:34 IST" duration-ms="51516">
+  <suite started-at="2025-05-08T21:27:26 IST" name="Default suite" finished-at="2025-05-08T21:27:58 IST" duration-ms="31311">
     <groups>
     </groups>
-    <test started-at="2025-05-08T19:56:43 IST" name="Default test" finished-at="2025-05-08T19:57:34 IST" duration-ms="51516">
+    <test started-at="2025-05-08T21:27:26 IST" name="Default test" finished-at="2025-05-08T21:27:58 IST" duration-ms="31311">
       <class name="tests.LoginTest">
-        <test-method is-config="true" signature="setUp()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:56:43 IST" name="setUp" finished-at="2025-05-08T19:57:16 IST" duration-ms="33508" status="PASS">
+        <test-method is-config="true" signature="setUp()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:27 IST" name="setUp" finished-at="2025-05-08T21:27:40 IST" duration-ms="13601" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- setUp -->
-        <test-method signature="LoginFunc(java.lang.String,java.lang.String)[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:16 IST" name="LoginFunc" data-provider="userCredentialsFromInternalArray" finished-at="2025-05-08T19:57:19 IST" duration-ms="2749" status="PASS">
+        <test-method signature="LoginFunc(java.lang.String,java.lang.String)[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:40 IST" name="LoginFunc" data-provider="userCredentialsFromInternalArray" finished-at="2025-05-08T21:27:43 IST" duration-ms="2460" status="PASS">
           <params>
             <param index="0">
               <value>
@@ -27,15 +27,15 @@
           <reporter-output>
           </reporter-output>
         </test-method> <!-- LoginFunc -->
-        <test-method is-config="true" signature="tearDown()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:19 IST" name="tearDown" finished-at="2025-05-08T19:57:20 IST" duration-ms="645" status="PASS">
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:43 IST" name="tearDown" finished-at="2025-05-08T21:27:44 IST" duration-ms="1053" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- tearDown -->
-        <test-method is-config="true" signature="setUp()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:20 IST" name="setUp" finished-at="2025-05-08T19:57:24 IST" duration-ms="4434" status="PASS">
+        <test-method is-config="true" signature="setUp()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:44 IST" name="setUp" finished-at="2025-05-08T21:27:47 IST" duration-ms="3236" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- setUp -->
-        <test-method signature="LoginFunc(java.lang.String,java.lang.String)[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:24 IST" name="LoginFunc" data-provider="userCredentialsFromInternalArray" finished-at="2025-05-08T19:57:26 IST" duration-ms="1420" status="PASS">
+        <test-method signature="LoginFunc(java.lang.String,java.lang.String)[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:47 IST" name="LoginFunc" data-provider="userCredentialsFromInternalArray" finished-at="2025-05-08T21:27:49 IST" duration-ms="1895" status="PASS">
           <params>
             <param index="0">
               <value>
@@ -51,15 +51,15 @@
           <reporter-output>
           </reporter-output>
         </test-method> <!-- LoginFunc -->
-        <test-method is-config="true" signature="tearDown()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:26 IST" name="tearDown" finished-at="2025-05-08T19:57:26 IST" duration-ms="564" status="PASS">
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:49 IST" name="tearDown" finished-at="2025-05-08T21:27:49 IST" duration-ms="356" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- tearDown -->
-        <test-method is-config="true" signature="setUp()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:26 IST" name="setUp" finished-at="2025-05-08T19:57:29 IST" duration-ms="2231" status="PASS">
+        <test-method is-config="true" signature="setUp()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:49 IST" name="setUp" finished-at="2025-05-08T21:27:51 IST" duration-ms="1599" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- setUp -->
-        <test-method signature="LoginFunc(java.lang.String,java.lang.String)[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:29 IST" name="LoginFunc" data-provider="userCredentialsFromInternalArray" finished-at="2025-05-08T19:57:33 IST" duration-ms="4535" status="PASS">
+        <test-method signature="LoginFunc(java.lang.String,java.lang.String)[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:51 IST" name="LoginFunc" data-provider="userCredentialsFromInternalArray" finished-at="2025-05-08T21:27:57 IST" duration-ms="6315" status="PASS">
           <params>
             <param index="0">
               <value>
@@ -75,7 +75,7 @@
           <reporter-output>
           </reporter-output>
         </test-method> <!-- LoginFunc -->
-        <test-method is-config="true" signature="tearDown()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T19:57:33 IST" name="tearDown" finished-at="2025-05-08T19:57:34 IST" duration-ms="501" status="PASS">
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:tests.LoginTest@1e800aaa]" started-at="2025-05-08T21:27:57 IST" name="tearDown" finished-at="2025-05-08T21:27:58 IST" duration-ms="572" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- tearDown -->
